separate hh reorg model utils
